Prueba_peigo/requirements.txt
# Dependencias principales
requests>=2.28.0
python-dotenv>=0.20.0
pyfiglet>=0.8.post1
python-pptx>=0.6.21

# Para visualizaciones
matplotlib>=3.5.0
numpy>=1.20.0

# Para testing
pytest>=7.0.0
